Rename login_page_3.py to login_page (second version).py
in this version, i asthetically changed the design of the login page from the original to the design that i made during the beggining of my project. However this version consists of a bug where the login button in not connceted to the login function meaning that there is no command being called when the login button is pressed.